Añadir ofertas arreglado. Implementado ver ofertas publicadas y compradas. Implementado eliminar ofertas publicadas

+++ sdi-entrega1-711/src/main/java/com/uniovi/service/OffersService.java
@@ -32,6 +32,12 @@ public class OffersService {
 	}
 
 	public void deleteOffer(Long id) {
+		User usuarioActual=userService.usuarioActual();
+		Offer oferta=getOffer(id);
+		if(oferta==null || !oferta.getUser().equals(usuarioActual))
+		{
+			throw new IllegalArgumentException();
+		}
 		offersRepository.deleteById(id);
 	}
